Completed LinearINodeManager set method

+++ src/lib/Block.h
@@ -3,7 +3,7 @@
 #include <cstdint>
 
 struct Block {
-  static const uint64_t BLOCK_SIZE = 4096;
+  static const uint16_t BLOCK_SIZE = 4096;
   typedef uint64_t ID;
 
   char data[BLOCK_SIZE];
